pyproject.toml
[build-system]
requires = ["setuptools >= 68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edgar-sentinel"
version = "0.1.0"
description = "Trading signal generator from SEC EDGAR filing sentiment analysis"
readme = "README.md"
license = "MIT"
requires-python = ">= 3.11"
authors = [
    { name = "Agent" }
]
keywords = ["sec", "edgar", "sentiment", "trading", "signals", "nlp", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]

dependencies = [
    # Core
    "pydantic >= 2.0",
    "pyyaml >= 6.0",
    # Ingestion
    "httpx >= 0.25",
    "beautifulsoup4 >= 4.12",
    "lxml >= 4.9",
    # Analyzers
    "scikit-learn >= 1.3",
    "numpy >= 1.24",
    "pandas >= 2.0",
    # CLI
    "click >= 8.0",
    "rich >= 13.0",
]

[project.optional-dependencies]
api = [
    "fastapi >= 0.100",
    "uvicorn[standard] >= 0.23",
]
llm = [
    "anthropic >= 0.30",
    "openai >= 1.0",
]
backtest = [
    "yfinance >= 0.2.30",
    "statsmodels >= 0.14",
]
all = [
    "edgar-sentinel[api,llm,backtest]",
]
dev = [
    "edgar-sentinel[all]",
    "pytest >= 8.0",
    "pytest-cov >= 4.0",
    "pytest-asyncio >= 0.23",
    "pytest-httpx >= 0.30",
    "respx >= 0.21",
    "ruff >= 0.3",
    "mypy >= 1.8",
]

[project.scripts]
edgar-sentinel = "edgar_sentinel.cli:main"

[project.urls]
Homepage = "https://github.com/claudiusbirdwhistle/edgar-sentinel"
Repository = "https://github.com/claudiusbirdwhistle/edgar-sentinel"
Issues = "https://github.com/claudiusbirdwhistle/edgar-sentinel/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no I/O, no network)",
    "integration: Integration tests (local I/O, no network)",
    "e2e: End-to-end tests (requires network access to EDGAR)",
]
addopts = "-m 'not e2e' --strict-markers"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
